refactor: rename Etats folder to states
