+++ src/index.ts
@@ -1,3 +1,3 @@
 require('pixi');
 require('p2');
-require('./Etats/game');
+require('./states/game');
